backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.31.0
alembic==1.13.2
greenlet==3.1.0

# Redis and Celery
celery[redis]==5.4.0
redis==5.1.1

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# HTTP client
httpx==0.27.2

# Web scraping
beautifulsoup4==4.12.3
trafilatura==1.12.2
feedparser==6.0.11
lxml==5.3.0

# Telegram
python-telegram-bot==21.6

# AI APIs
anthropic>=0.84.0
google-genai>=1.0.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
